requirements.txt
faiss-cpu>=1.8.0
numpy>=1.26.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
rank-bm25>=0.2.2
requests>=2.31.0
sentence-transformers>=3.0.0
